% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{computeInclusionList}
\alias{computeInclusionList}
\title{Compute inclusion list}
\usage{
computeInclusionList(
  query_sample,
  jsd_ranks,
  background,
  z_score_cutoff = -1.5
)
}
\arguments{
\item{query_sample}{A one-column data frame containing gene expression values. Rownames have to be gene
identifiers corresponding to the background data frame.}

\item{jsd_ranks}{The ranked Jennsen-Shannon-Divergence values for all samples in the background.}

\item{background}{A data frame containing gene expression values. Rownames have to be gene
identifiers corresponding to the query sample.}

\item{z_score_cutoff}{A Z-score below which a gene is considered significant. (DEFAULT: -1.5)}
}
\value{
A vector of genes considered significantly unique.
}
\description{
Identifies a set of genes whose Jennsen-Shannon-Divergence value could be considered
significant.
}
