% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{computeJSDBackground}
\alias{computeJSDBackground}
\title{Compute Jennsen-Shannon-Divergence for all samples}
\usage{
computeJSDBackground(background)
}
\arguments{
\item{background}{A data frame containing gene expression values. Rownames have to be gene
identifiers corresponding to the query sample.}
}
\value{
A data frame of Jennsen-Shannon-Divergence values and their ranks with respect to all samples.
}
\description{
Compute Jennsen-Shannon-Divergence for a all samples in a set of background
samples.
}
