% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{computeJSDSingle}
\alias{computeJSDSingle}
\title{Compute Jennsen-Shannon-Divergence for a single sample}
\usage{
computeJSDSingle(querySample, background)
}
\arguments{
\item{querySample}{A one-column data frame containing gene expression values. Rownames have to be gene
identifiers corresponding to the background data frame.}

\item{background}{A data frame containing gene expression values. Rownames have to be gene
identifiers corresponding to the query sample.}
}
\value{
The value of the Jennsen-Shannon-Divergence.
}
\description{
Compute Jennsen-Shannon-Divergence for a single sample given a data frame of background
samples.
}
