% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBackground.R
\name{createThresholdDist}
\alias{createThresholdDist}
\title{Create Threshold Distribution}
\usage{
createThresholdDist(gexp, numBootstrapSamples = 1000)
}
\arguments{
\item{gexp}{A vector of gene expression values.}

\item{numBootstrapSamples}{Number of gene expression samples to draw (with replacement). (DEFAULT: 1000)}
}
\value{
A data frame containing lower and upper gene expression thresholds.
}
\description{
Computes a distribution of lower and upper gene expression thresholds of a single gene.
}
