% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBackground.R
\name{.optLowFun}
\alias{.optLowFun}
\title{Helper function}
\usage{
.optLowFun(x, empcdf, maxVal)
}
\arguments{
\item{x}{A gene expression value.}

\item{empcdf}{An empirical cumulative distribution function.}

\item{maxVal}{The maximum expression value to consider. (UNUSED)}
}
\value{
The size of the rectangle.
}
\description{
Computes the size of a rectangle above a given empirical cumulative distribution function.
}
