% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{generateNetwork}
\alias{generateNetwork}
\title{Generate networks from .bed files}
\usage{
generateNetwork(
  prom_df,
  enh_df,
  exprTFs_df,
  Dnase_df,
  tfs_df,
  Cistrome_ChIPseq_sorted_df,
  Processed_GeneHancer_df,
  KnownPromoters_df
)
}
\arguments{
\item{prom_df}{A data frame in .bed format. Each line must be a H3K4me3 peak.}

\item{enh_df}{A data frame in .bed format. Each line must be a H3K27ac peak.}

\item{exprTFs_df}{A one column data frame of all expressed genes.}

\item{Dnase_df}{A data frame in .bed format. Each line must be a DNase-seq peak}

\item{tfs_df}{A one column data frame of core TFs to consider.}

\item{Cistrome_ChIPseq_sorted_df}{A data frame of TF ChIP-seq binding sites. Columns must correspond to:
Chromosome, Start, End, TF (in this order).}

\item{Processed_GeneHancer_df}{Data frame of the processed gene hancer data.}

\item{KnownPromoters_df}{Data frame of known promoter regions (from AnimalTFDB).}
}
\value{
A list of data frames corresponding to TF binding sites in active promoters and enhancers for core
TFs and non-core mutually connected neighbors.
}
\description{
Puts together the gene regulatory rules of all genes to be considered in a network
}
