% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{generateNetworkLogic}
\alias{generateNetworkLogic}
\title{Generate Network Logics}
\usage{
generateNetworkLogic(data_prom, data_enh, ppi, isStrict = F)
}
\arguments{
\item{data_prom}{A data frame with TF binding sites in (active) promoter regions.}

\item{data_enh}{A data frame with TF binding sites in (active) enhancer regions.}

\item{ppi}{A data frame of protein-protein interactions.}

\item{isStrict}{Should weakly or strongly connected componentes of overlapping binding
sites be computed?}
}
\value{
A list containing only the promoter/enhancer network rules, the combined network rules
and the unique regulators per gene in enhancers and promoters.
}
\description{
Puts together the gene regulatory rules of all genes to be considered in a network
}
