% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{getAllRegulators}
\alias{getAllRegulators}
\title{Compute Regulators of a gene}
\usage{
getAllRegulators(network_df)
}
\arguments{
\item{network_df}{A two-column data frame with the first column containing genes and the second
column containing the regulatory rules.}
}
\value{
A data frame of unique regulators per gene.
}
\description{
Given a data frame of gene regulatory logic rules, computes all regulators.
}
