% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{normalizeExpression}
\alias{normalizeExpression}
\title{Normalize expression matrix}
\usage{
normalizeExpression(exp_df)
}
\arguments{
\item{exp_df}{A matrix/data frame of gene expression values with rows corresponding to genes
and columns to samples.}
}
\value{
A matrix/data frame with normalized expression values.
}
\description{
Normalizes each gene (row) by its maximum value.
}
