% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBackground.R
\name{tpm3}
\alias{tpm3}
\title{Counts to TPM}
\usage{
tpm3(counts, len)
}
\arguments{
\item{counts}{A matrix or data frame object with gene expression counts. Columns correspond to samples and
rows to genes.}

\item{len}{A vector of gene lengths having the same number of entries as there are rows in **counts**,}
}
\value{
A matrix/data frame object with TPM values having the same dimensions as **counts**.
}
\description{
Transform a matrix/data frame of gene expression counts into a matrix/data frame of
TPM values.
}
