% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{writeLogic}
\alias{writeLogic}
\title{Compute logic rules of genes}
\usage{
writeLogic(df, ppi, directed, mode)
}
\arguments{
\item{df}{A data frame with TF binding sites in regulatory regions of genes.}

\item{ppi}{A data frame of protein-protein interactions.}

\item{directed}{Should the network of overlapping binding sites be considered directed or undirected?}

\item{mode}{Either "weak" or "strong". Should weakly or strongly connected componentes of overlapping binding
sites be computed?}
}
\value{
A data frame of gene regulatory logic rules for each gene.
}
\description{
Puts together the gene regulatory rules of all genes to be considered in a network
}
