% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworks.R
\name{ztest}
\alias{ztest}
\title{Compute Z-scores}
\usage{
ztest(x, m, s)
}
\arguments{
\item{x}{A numeric value or vector of numeric values to standardize.}

\item{m}{A numeric value or vector of numeric values corresponding to the distribution mean(s).}

\item{s}{A numeric value or vector of numeric values corresponding to the distribution standard deviation(s).}
}
\value{
A numeric value or vector of numeric values of Z-scores.
}
\description{
Computes the Z-score of values given the distribution means and standard deviations.
}
